% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AF}
\alias{AF}
\title{CD34+ count cell data}
\format{
A data frame with 40 rows and 3 variables:
\describe{
\item{y}{CD34+ counts}
\item{id}{Subject}
\item{met}{Method}
}
}
\usage{
AF
}
\description{
A dataset containing CD34+ count cell data obtained by two methods
}
\references{
{
Fornas, O., Garcia, J., and Petriz, J. (2000). Flow cytometry counting of CD34+ cells in whole blood. Nature Medicine 6, 833–836.
}
}
\keyword{datasets}
