% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPP}
\alias{EPP}
\title{Sparrow fledglings paternity data}
\format{
A data frame with 193 rows and 3 variables:
\describe{
\item{Social}{Fledglings counts}
\item{id}{Subject identifier}
\item{Year}{Breeding season}
}
}
\usage{
EPP
}
\description{
A dataset containing the counts of fledglings a male had in every breeding season
}
\references{
{
Schroeder, J., Burke, T., Mannarelli, M. E., Dawson, D. A., & Nakagawa, S. (2012). Maternal effects and heritability of annual productivity. Journal of Evolutionary Biology, 25, 149– 156.
}
}
\keyword{datasets}
