% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{fit_model_rep}
\alias{fit_model_rep}
\title{Estimates GLMM. Repeatability setting.}
\usage{
fit_model_rep(
  data,
  fam = c("poisson", "nbinom2", "nbinom1", "zip", "zinb1", "zinb2")
)
}
\arguments{
\item{data}{A data frame containing at least two columns: outcome (named as y) and subject identifier (named as id).}

\item{fam}{Character string. The within-subjects pdf to use. Valid options are: "poisson" (default) for Poisson pdf; "nbinom1" for Negative Binomial pdf with additive extradispersion; "nbinom2" for Negative Binomial pdf with proportional extradispersion; "zip" for zero-inflated Poisson pdf; "zinb1" for zero-inflated nbinom1 pdf; "zinb2" for zero-inflated nbinom2 pdf;}
}
\value{
An object of class glmmTMB with the model estimates.
}
\description{
Estimates the GLMM model for concordance setting
}
\keyword{internal}
