% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grimso}
\alias{Grimso}
\title{Tick counts data}
\format{
A data frame with 520 rows and 3 variables:
\describe{
\item{Tot}{Tick counts}
\item{TransectID}{Subject identifier}
\item{Round}{Identifier of repeated measurements}
}
}
\usage{
Grimso
}
\description{
A dataset containing the counts of ticks in Grimsö area
}
\references{
{
Kjellander, P.L., Aronsson, M., Bergvall, U.A. et al. (2021). Validating a common tick survey method: cloth-dragging and line transects. Exp Appl Acarol 83, 131–146.
}
}
\keyword{datasets}
