% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_est_rep_zi}
\alias{r_est_rep_zi}
\title{ICC repeatability for zero-inflated models}
\usage{
r_est_rep_zi(data, model)
}
\arguments{
\item{data}{A data frame containing at least two columns: outcome (named as y) and subject identifier (named as id).}

\item{model}{A glmmTMB obejct.}
}
\value{
A data frame with ICC estimate, its standard error and variance components.
}
\description{
Estimates the ICC for repeatability setting for zero-inflated models (poisson, nbinom1 and nbinom2 families).
}
\keyword{internal}
