% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_NB1_2}
\alias{r_NB1_2}
\title{ICC for Negative Binomial with additive extradispersion. Concordance setting}
\usage{
r_NB1_2(b1, b2, b3, sb)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra-dispersion in the GLMM model}

\item{sb}{Between-methods variability}
}
\value{
Scalar
}
\description{
ICC for Negative Binomial with additive extradispersion. Concordance setting
}
\keyword{internal}
