% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_NB2}
\alias{r_NB2}
\title{ICC for Negative Binomial with proportional extradispersion. Repeatability setting}
\usage{
r_NB2(b1, b2, b3)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra-dispersion in the GLMM model}
}
\value{
Scalar
}
\description{
ICC for Negative Binomial with proportional extradispersion. Repeatability setting
}
\keyword{internal}
