% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_Pois}
\alias{r_Pois}
\title{ICC Poisson model. Repeatability setting}
\usage{
r_Pois(b1, b2)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}
}
\value{
Scalar
}
\description{
ICC for Poisson model. Repeatability setting
}
\keyword{internal}
