% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_Pois2}
\alias{r_Pois2}
\title{ICC Poisson model. Concordance setting}
\usage{
r_Pois2(b1, b2, sb)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{sb}{Between-methods variability}
}
\value{
Scalar
}
\description{
ICC for Poisson model. Concordance setting
}
\keyword{internal}
