% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_ZIP}
\alias{r_ZIP}
\title{ICC Zero-Inflated Poisson model. Repeatability setting}
\usage{
r_ZIP(b1, b2, b3)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra zero probability in the GLMM model}
}
\value{
Scalar
}
\description{
ICC for Zero-Inflated Poisson model. Repeatability setting
}
\keyword{internal}
