% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{r_ZIP2}
\alias{r_ZIP2}
\title{ICC Zero-Inflated Poisson model. Concordance setting}
\usage{
r_ZIP2(b1, b2, b3, sb)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra zero probability in the GLMM model}

\item{sb}{Between-methods variability}
}
\value{
Scalar
}
\description{
ICC for Zero-Inflated Poisson model. Concordance setting
}
\keyword{internal}
