% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr}
\alias{rqr}
\title{RQR for GLMM
Computes the randomized quantile residuals for GLMM}
\usage{
rqr(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for GLMM
Computes the randomized quantile residuals for GLMM
}
\details{
Randomized quantile residuals (RQR) are computed for GLMMM. The within-cluster families considered
are Poisson, Negative Binomial with additive and proportional extra-dispersion and their zero-inflated extensions.
For further details on RQR see Dunn and Smyth (1996) and Feng et al (2020)
}
\references{
{

Dunn PK, Smyth GK. (1996). Randomized quantile residuals. J Comput Graph Stat.5(3):236–44.

Feng et al. (2020). A comparison of residual diagnosis tools for diagnosing regression models for count data. BMC Medical Research Methodology 20:175
}
}
\keyword{internal}
