% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_nb1}
\alias{rqr_nb1}
\title{RQR for NegBin1 GLMM
Computes the randomized quantile residuals for NegBin1 GLMM}
\usage{
rqr_nb1(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for NegBin1 GLMM
Computes the randomized quantile residuals for NegBin1 GLMM
}
\keyword{internal}
