% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_nb2}
\alias{rqr_nb2}
\title{RQR for NegBin2 GLMM
Computes the randomized quantile residuals for NegBin2 GLMM}
\usage{
rqr_nb2(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for NegBin2 GLMM
Computes the randomized quantile residuals for NegBin2 GLMM
}
\keyword{internal}
