% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_pois}
\alias{rqr_pois}
\title{RQR for Poisson GLMM
Computes the randomized quantile residuals for Poisson GLMM}
\usage{
rqr_pois(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for Poisson GLMM
Computes the randomized quantile residuals for Poisson GLMM
}
\keyword{internal}
