% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_zinb1}
\alias{rqr_zinb1}
\title{RQR for ZINB1 GLMM
Computes the randomized quantile residuals for ZINB1 GLMM}
\usage{
rqr_zinb1(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for ZINB1 GLMM
Computes the randomized quantile residuals for ZINB1 GLMM
}
\keyword{internal}
