% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_zinb2}
\alias{rqr_zinb2}
\title{RQR for ZINB2 GLMM
Computes the randomized quantile residuals for ZINB2 GLMM}
\usage{
rqr_zinb2(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for ZINB2 GLMM
Computes the randomized quantile residuals for ZINB2 GLMM
}
\keyword{internal}
