% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_functions.R
\name{rqr_zip}
\alias{rqr_zip}
\title{RQR for ZIP GLMM
Computes the randomized quantile residuals for ZIP GLMM}
\usage{
rqr_zip(x)
}
\arguments{
\item{x}{An object of clas \emph{iccc}.}
}
\value{
A vector with the residuals.
}
\description{
RQR for ZIP GLMM
Computes the randomized quantile residuals for ZIP GLMM
}
\keyword{internal}
