% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{se_r_Pois}
\alias{se_r_Pois}
\title{ICC standard error for Poisson model. Repeatability setting}
\usage{
se_r_Pois(b1, b2, CovB)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{CovB}{Variance covariance matrix of b1 and b2}
}
\value{
Scalar
}
\description{
ICC standard error for Poisson model. Repeatability setting
}
\keyword{internal}
