% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{se_r_Pois2}
\alias{se_r_Pois2}
\title{ICC standard error for Poisson model. Concordance setting}
\usage{
se_r_Pois2(b1, b2, sb, CovB, var.sb)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{sb}{Between-methods variability}

\item{CovB}{Variance covariance matrix of b1 and b2}

\item{var.sb}{Standard error of sb (in variance scale).}
}
\value{
Scalar
}
\description{
ICC standard error for Poisson model. Concordance setting
}
\keyword{internal}
