% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{se_r_ZINB1}
\alias{se_r_ZINB1}
\title{ICC standard error for Zero-Inflated Negative Binomial with a additive extradispersion. Repeatability setting}
\usage{
se_r_ZINB1(b1, b2, b3, b4, CovB)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra-dispersion in the GLMM model}

\item{b4}{Parameter related to extra zero probability in the GLMM model}

\item{CovB}{Variance covariance matrix of b1, b2, b3 and b4}
}
\value{
Scalar
}
\description{
ICC standard error for Zero-Inflated Negative Binomial with additive extradispersion. Repeatability setting
}
\keyword{internal}
