% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{se_r_ZIP}
\alias{se_r_ZIP}
\title{ICC standard error for Zero-Inflated Poisson model. Repeatability setting}
\usage{
se_r_ZIP(b1, b2, b3, CovB)
}
\arguments{
\item{b1}{Intercept of GLMM}

\item{b2}{Parameter related to between-subjects variance in the GLMM model}

\item{b3}{Parameter related to extra zero probability in the GLMM model}

\item{CovB}{Variance covariance matrix of b1,b2 and b3}
}
\value{
Scalar
}
\description{
ICC standard error for Zero-Inflated Poisson model. Repeatability setting
}
\keyword{internal}
